Trekker ut vilkårbegrunnelser i egen hook.
