+++ src/frontend/komponenter/Fagsak/Vedtak/VedtakBegrunnelserTabell/Context/VedtaksperiodeMedBegrunnelserContext.ts
@@ -1,28 +1,13 @@
-import React, { useEffect } from 'react';
-
 import constate from 'constate';
 
-import { ActionMeta, GroupType, ISelectOption } from '@navikt/familie-form-elements';
-import { useHttp } from '@navikt/familie-http';
+import { ActionMeta, ISelectOption } from '@navikt/familie-form-elements';
 import { FeltState, ok, useFelt, useSkjema } from '@navikt/familie-skjema';
-import { Ressurs, byggTomRessurs, RessursStatus } from '@navikt/familie-typer';
 
 import { IBehandling } from '../../../../../typer/behandling';
 import { IFagsak } from '../../../../../typer/fagsak';
-import {
-    IRestVedtakBegrunnelseTilknyttetVilkår,
-    VedtakBegrunnelse,
-    VedtakBegrunnelseType,
-    vedtakBegrunnelseTyper,
-} from '../../../../../typer/vedtak';
-import {
-    IRestVedtaksbegrunnelse,
-    IVedtaksperiodeMedBegrunnelser,
-    Vedtaksperiodetype,
-} from '../../../../../typer/vedtaksperiode';
-import { Vilkårsbegrunnelser, VilkårType } from '../../../../../typer/vilkår';
+import { IVedtaksperiodeMedBegrunnelser } from '../../../../../typer/vedtaksperiode';
 import { IPeriode } from '../../../../../utils/kalender';
-import { hentUtgjørendeVilkårImpl } from '../Hooks/useVedtakBegrunnelseMultiselect';
+import { useVilkårBegrunnelser } from '../Hooks/useVilkårBegrunnelser';
 
 interface IProps {
     fagsak: IFagsak;
@@ -32,8 +17,6 @@ interface IProps {
 
 const [VedtaksperiodeMedBegrunnelserProvider, useVedtaksperiodeMedBegrunnelser] = constate(
     ({ åpenBehandling, vedtaksperiodeMedBegrunnelser }: IProps) => {
-        const { request } = useHttp();
-
         const periode = useFelt<IPeriode>({
             verdi: {
                 fom: vedtaksperiodeMedBegrunnelser.fom,
@@ -51,10 +34,6 @@ const [VedtaksperiodeMedBegrunnelserProvider, useVedtaksperiodeMedBegrunnelser] 
             valideringsfunksjon: (felt: FeltState<ISelectOption[]>) => ok(felt),
         });
 
-        const [vilkårBegrunnelser, settVilkårbegrunnelser] = React.useState<
-            Ressurs<Vilkårsbegrunnelser>
-        >(byggTomRessurs());
-
         const { skjema } = useSkjema<
             {
                 periode: IPeriode;
@@ -71,25 +50,12 @@ const [VedtaksperiodeMedBegrunnelserProvider, useVedtaksperiodeMedBegrunnelser] 
             skjemanavn: 'Begrunnelser for vedtaksperiode',
         });
 
-        useEffect(() => {
-            hentVilkårBegrunnelseTekster();
-        }, []);
-
-        useEffect(() => {
-            begrunnelser.validerOgSettFelt(
-                mapBegrunnelserTilSelectOptions(vedtaksperiodeMedBegrunnelser, vilkårBegrunnelser)
-            );
-        }, [vilkårBegrunnelser]);
-
-        const hentVilkårBegrunnelseTekster = () => {
-            request<void, Vilkårsbegrunnelser>({
-                method: 'GET',
-                url: `/familie-ba-sak/api/vilkaarsvurdering/vilkaarsbegrunnelser`,
-                påvirkerSystemLaster: true,
-            }).then((vilkårBegrunnelser: Ressurs<Vilkårsbegrunnelser>) => {
-                settVilkårbegrunnelser(vilkårBegrunnelser);
-            });
-        };
+        const { grupperteBegrunnelser, vilkårBegrunnelser } = useVilkårBegrunnelser({
+            åpenBehandling,
+            vedtaksperiodeMedBegrunnelser,
+            periode: skjema.felter.periode.verdi,
+            begrunnelser,
+        });
 
         const onChangeBegrunnelse = (action: ActionMeta<ISelectOption>) => {
             if (action.option)
@@ -141,74 +107,6 @@ const [VedtaksperiodeMedBegrunnelserProvider, useVedtaksperiodeMedBegrunnelser] 
                 }
         };
 
-        const vedtaksperiodeTilVedtakBegrunnelseTyper = () => {
-            switch (vedtaksperiodeMedBegrunnelser.type) {
-                case Vedtaksperiodetype.UTBETALING:
-                    return [VedtakBegrunnelseType.INNVILGELSE, VedtakBegrunnelseType.REDUKSJON];
-                case Vedtaksperiodetype.FORTSATT_INNVILGET:
-                    return [VedtakBegrunnelseType.FORTSATT_INNVILGET];
-                case Vedtaksperiodetype.OPPHØR:
-                    return [VedtakBegrunnelseType.OPPHØR];
-                default:
-                    return [];
-            }
-        };
-
-        const vedtakBegrunnelseTyperKnyttetTilVedtaksperiodetype = vedtaksperiodeTilVedtakBegrunnelseTyper();
-
-        const hentUtgjørendeVilkår = (begrunnelseType: VedtakBegrunnelseType): VilkårType[] =>
-            hentUtgjørendeVilkårImpl(
-                begrunnelseType,
-                åpenBehandling.personResultater,
-                skjema.felter.periode.verdi
-            );
-
-        const grupperteBegrunnelser =
-            vilkårBegrunnelser.status === RessursStatus.SUKSESS
-                ? Object.keys(vilkårBegrunnelser.data)
-                      .filter((vedtakBegrunnelseType: string) =>
-                          vedtakBegrunnelseTyperKnyttetTilVedtaksperiodetype.includes(
-                              vedtakBegrunnelseType as VedtakBegrunnelseType
-                          )
-                      )
-                      .reduce((acc: GroupType<ISelectOption>[], vedtakBegrunnelseType: string) => {
-                          const utgjørendeVilkårForPeriodeOgResultat: VilkårType[] = hentUtgjørendeVilkår(
-                              vedtakBegrunnelseType as VedtakBegrunnelseType
-                          );
-                          return [
-                              ...acc,
-                              {
-                                  label:
-                                      vedtakBegrunnelseTyper[
-                                          vedtakBegrunnelseType as VedtakBegrunnelseType
-                                      ],
-                                  options: vilkårBegrunnelser.data[
-                                      vedtakBegrunnelseType as VedtakBegrunnelseType
-                                  ]
-                                      .filter(
-                                          (
-                                              restVedtakBegrunnelseTilknyttetVilkår: IRestVedtakBegrunnelseTilknyttetVilkår
-                                          ) => {
-                                              return restVedtakBegrunnelseTilknyttetVilkår.vilkår
-                                                  ? utgjørendeVilkårForPeriodeOgResultat.includes(
-                                                        restVedtakBegrunnelseTilknyttetVilkår.vilkår
-                                                    )
-                                                  : false;
-                                          }
-                                      )
-                                      .map(
-                                          (
-                                              restVedtakBegrunnelseTilknyttetVilkår: IRestVedtakBegrunnelseTilknyttetVilkår
-                                          ) => ({
-                                              label: restVedtakBegrunnelseTilknyttetVilkår.navn,
-                                              value: restVedtakBegrunnelseTilknyttetVilkår.id,
-                                          })
-                                      ),
-                              },
-                          ];
-                      }, [])
-                : [];
-
         return {
             id: vedtaksperiodeMedBegrunnelser.id,
             skjema,
@@ -218,42 +116,5 @@ const [VedtaksperiodeMedBegrunnelserProvider, useVedtaksperiodeMedBegrunnelser] 
         };
     }
 );
-
-const mapBegrunnelserTilSelectOptions = (
-    vedtaksperiodeMedBegrunnelser: IVedtaksperiodeMedBegrunnelser,
-    vilkårBegrunnelser: Ressurs<Vilkårsbegrunnelser>
-): ISelectOption[] => {
-    return vedtaksperiodeMedBegrunnelser.begrunnelser
-        .filter(
-            (begrunnelse: IRestVedtaksbegrunnelse) =>
-                begrunnelse.vedtakBegrunnelseSpesifikasjon !==
-                    VedtakBegrunnelse.REDUKSJON_FRITEKST &&
-                begrunnelse.vedtakBegrunnelseSpesifikasjon !== VedtakBegrunnelse.AVSLAG_FRITEKST &&
-                begrunnelse.vedtakBegrunnelseSpesifikasjon !== VedtakBegrunnelse.OPPHØR_FRITEKST &&
-                begrunnelse.vedtakBegrunnelseSpesifikasjon !==
-                    VedtakBegrunnelse.FORTSATT_INNVILGET_FRITEKST
-        )
-        .map((begrunnelse: IRestVedtaksbegrunnelse) => ({
-            value: begrunnelse.vedtakBegrunnelseSpesifikasjon.toString(),
-            label: hentLabelForOption(
-                begrunnelse.vedtakBegrunnelseType,
-                begrunnelse.vedtakBegrunnelseSpesifikasjon,
-                vilkårBegrunnelser
-            ),
-        }));
-};
-
-const hentLabelForOption = (
-    vedtakBegrunnelseType: VedtakBegrunnelseType,
-    vedtakBegrunnelseSpesifikasjon: VedtakBegrunnelse,
-    vilkårBegrunnelser: Ressurs<Vilkårsbegrunnelser>
-) => {
-    return vilkårBegrunnelser.status === RessursStatus.SUKSESS
-        ? vilkårBegrunnelser.data[vedtakBegrunnelseType].find(
-              (restVedtakBegrunnelseTilknyttetVilkår: IRestVedtakBegrunnelseTilknyttetVilkår) =>
-                  restVedtakBegrunnelseTilknyttetVilkår.id === vedtakBegrunnelseSpesifikasjon
-          )?.navn ?? ''
-        : '';
-};
 
 export { VedtaksperiodeMedBegrunnelserProvider, useVedtaksperiodeMedBegrunnelser };
